Trait peresil::Recoverable

pub trait Recoverable {
    fn recoverable(&self) -> bool;
}

Indicate if an error should terminate all parsing.

Non-recoverable errors will not allow for alternatives to be tried, basically unwinding the parsing stack all the way back to the beginning. Unrecoverable errors are useful for errors that indicate that the content was well-formed but not semantically correct.

Required Methods

fn recoverable(&self) -> bool